Migrate from Next.js App Router to Pages Router
