--- refine-backoffice/src/app/page.tsx
@@ -1,21 +1,0 @@
-'use client';
-
-import { useEffect } from 'react';
-import { useRouter } from 'next/navigation';
-
-export default function Home() {
-  const router = useRouter();
-
-  useEffect(() => {
-    router.push('/dashboard');
-  }, [router]);
-
-  return (
-    <div className="flex min-h-screen items-center justify-center">
-      <div className="text-center">
-        <h1 className="text-2xl font-bold">Redirecting to Dashboard...</h1>
-        <p className="text-muted-foreground mt-2">Please wait...</p>
-      </div>
-    </div>
-  );
-}
