+++ refine-backoffice/src/pages/_app.tsx
@@ -1,0 +1,11 @@
+import type { AppProps } from 'next/app';
+import I18nProvider from '@/lib/i18n';
+import '@/app/globals.css';
+
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <I18nProvider>
+      <Component {...pageProps} />
+    </I18nProvider>
+  );
+}
